+++ eva/src/components/dashPanelBoard.vue
@@ -256,7 +256,11 @@
               :class="{ loading: search.status === 'pending' }"
             />
           </div>
-          <v-tooltip bottom :color="theme.$accent_ui_color">
+          <v-tooltip
+            z-index="99"
+            bottom
+            :color="theme.$accent_ui_color"
+          >
             <template v-slot:activator="{ on }">
               <v-icon
                 class="search-play"
@@ -269,7 +273,11 @@
             </template>
             <span>Запустить ИД</span>
           </v-tooltip>
-          <v-tooltip bottom :color="theme.$accent_ui_color">
+          <v-tooltip
+            z-index="99"
+            bottom
+            :color="theme.$accent_ui_color"
+          >
             <template v-slot:activator="{ on }">
               <v-icon
                 class="search-pencil"
@@ -282,7 +290,11 @@
             </template>
             <span>Редактировать ИД</span>
           </v-tooltip>
-          <v-tooltip bottom :color="theme.$accent_ui_color">
+          <v-tooltip
+            z-index="99"
+            bottom
+            :color="theme.$accent_ui_color"
+          >
             <template v-slot:activator="{ on }">
               <v-icon
                 class="search-clock"
@@ -295,7 +307,11 @@
             </template>
             <span>Планировщик</span>
           </v-tooltip>
-          <v-tooltip bottom :color="theme.$accent_ui_color">
+          <v-tooltip
+            z-index="99"
+            bottom
+            :color="theme.$accent_ui_color"
+          >
             <template v-slot:activator="{ on }">
               <v-icon
                 class="search-clock"
@@ -309,7 +325,11 @@
             </template>
             <span>Экспортировать ИД</span>
           </v-tooltip>
-          <v-tooltip bottom :color="theme.$accent_ui_color">
+          <v-tooltip
+            z-index="99"
+            bottom
+            :color="theme.$accent_ui_color"
+          >
             <template v-slot:activator="{ on }">
               <v-icon
                 class="search-clock"
@@ -322,7 +342,11 @@
             </template>
             <span>Создать отчет</span>
           </v-tooltip>
-          <v-tooltip bottom :color="theme.$accent_ui_color">
+          <v-tooltip
+            bottom
+            z-index="99"
+            :color="theme.$accent_ui_color"
+          >
             <template v-slot:activator="{ on }">
               <v-icon
                 class="search-trash"
@@ -769,34 +793,34 @@
 
 <script>
 import {
-  mdiPlusBox,
-  mdiDoor,
-  mdiCompare,
-  mdiScriptTextOutline,
-  mdiAnimationPlay,
-  mdiUndoVariant,
+  mdiAccount,
   mdiAccountEdit,
-  mdiPlay,
+  mdiAnimationPlay,
+  mdiArrowDownBold,
+  mdiCheckBold,
+  mdiClockOutline,
+  mdiCodeTags,
+  mdiCompare,
+  mdiContentSave,
+  mdiDatabase,
+  mdiDoor,
   mdiEye,
   mdiFileDocumentOutline,
-  mdiArrowDownBold,
-  mdiContentSave,
-  mdiAccount,
-  mdiHomeVariantOutline,
-  mdiSettings,
-  mdiHelpCircleOutline,
-  mdiClockOutline,
-  mdiDatabase,
-  mdiTableEdit,
-  mdiCodeTags,
-  mdiTrashCanOutline,
-  mdiMinusBox,
-  mdiToolbox,
-  mdiPencil,
-  mdiVariable,
-  mdiCheckBold,
-  mdiSwapVerticalBold,
   mdiFilter,
+  mdiHelpCircleOutline,
+  mdiHomeVariantOutline,
+  mdiMinusBox,
+  mdiPencil,
+  mdiPlay,
+  mdiPlusBox,
+  mdiScriptTextOutline,
+  mdiSettings,
+  mdiSwapVerticalBold,
+  mdiTableEdit,
+  mdiToolbox,
+  mdiTrashCanOutline,
+  mdiUndoVariant,
+  mdiVariable,
 } from '@mdi/js';
 import EvaLogo from '../images/eva-logo.svg';
 
@@ -828,7 +852,7 @@ export default {
       tocken_elem: false,
       profile_elem: false,
       save_elem: false,
-      editMode: process.env.VUE_APP_DASHBOARD_EDITING_MODE == 'true',
+      editMode: process.env.VUE_APP_DASHBOARD_EDITING_MODE === 'true',
       code_elem: false,
       check: mdiCheckBold,
       look: mdiEye,
@@ -946,7 +970,7 @@ export default {
     };
   },
   computed: {
-    idDash: function () {
+    idDash() {
       return this.idDashFrom;
     },
     headerTop() {
@@ -955,7 +979,7 @@ export default {
     isAdmin() {
       return this.userPermissions && this.userPermissions.includes('admin_all');
     },
-    searches: function () {
+    searches() {
       // массив со всеми ИС на странице
       let searchesRes = [];
       if (this.idDash) {
@@ -967,29 +991,25 @@ export default {
       }
       return searchesRes;
     },
-    name: function () {
+    name() {
       return this.$store.getters.getName(this.idDash);
     },
-    theme: function () {
+    theme() {
       return this.$store.getters.getTheme;
     },
-    editPermission: function () {
-      if (
-        this.userPermissions.includes('admin_all') ||
-        this.userPermissions.includes('editdash')
-      ) {
-        return true;
-      }
-      return false;
+    editPermission() {
+      return this.userPermissions.includes('admin_all') ||
+        this.userPermissions.includes('editdash');
+
     },
-    textareaEv: function () {
+    textareaEv() {
       let eventFull = this.$store.getters.getEventFull(this.idDash);
-      if (eventFull != '') {
-        this.textarea_event = eventFull;
+      if (eventFull !== '') {
+        this.setTextarea_event(eventFull);
       }
       return true;
     },
-    tockens: function () {
+    tockens() {
       // получения всех токенов на страницы
       let tockens = this.$store.getters.getTockens(this.idDash);
 
@@ -999,30 +1019,28 @@ export default {
       });
       return tockens;
     },
-    elements: function () {
+    elements() {
       // получение всех элемнета на странице
       return this.$store.getters.getElements(this.idDash);
     },
-    actions: function () {
+    actions() {
       // получение всех событий элемента на странице
       return function (element) {
-        let names = this.$store.getters
+        return this.$store.getters
           .getActions({ elem: element, idDash: this.idDash })
           .map((item) => {
             return item.name;
           });
-        return names;
       };
     },
-    capture: function () {
+    capture() {
       // получение всех подсобытий элемента на странице (события второго уровня )
       return function (element) {
-        let capture = this.$store.getters.getCapture({
+        return this.$store.getters.getCapture({
           elem: element.elem,
           action: element.action,
           idDash: this.idDash,
         });
-        return capture;
       };
     },
     blockToolStyle() {
@@ -1033,7 +1051,10 @@ export default {
       };
     },
     isEditDash() {
-      return !!(this.allGroups.find((group) => group.dashs.includes(this.name) && group.users.includes(this.login)));
+      return !!this.allGroups.find(
+        (group) =>
+          group.dashs.includes(this.name) && group.users.includes(this.login)
+      );
     },
   },
   mounted() {
@@ -1059,7 +1080,7 @@ export default {
       }
     };
     let eventFull = this.$store.getters.getEventFull(this.idDash);
-    if (eventFull != '') {
+    if (eventFull !== '') {
       this.textarea_event = eventFull;
     }
     if (document.querySelector('.block-code')) {
@@ -1079,6 +1100,9 @@ export default {
     window.removeEventListener('resize', this.updateScreenHeight);
   },
   methods: {
+    setTextarea_event(eventFull) {
+      this.textarea_event = eventFull;
+    },
     getGroups: function () {
       let response = this.$store.getters.getGroups();
       response.then((res) => {
@@ -1148,11 +1172,13 @@ export default {
         // я так понимаю если на странице есть созданные ИС
         Object.keys(this.change).forEach((item) => {
           // то пробегаемся по всем ИС
-          item == id ? (this.change[item] = true) : (this.change[item] = false); // если нашли выбронный ИС то меняем его статус
+          item === id
+            ? (this.change[item] = true)
+            : (this.change[item] = false); // если нашли выбронный ИС то меняем его статус
         });
         let search = this.searches.filter((item) => {
           // получаем только тот ИС который редактируется
-          return item.sid == id;
+          return item.sid === id;
         })[0];
         // отстутствие отступов сделано специально  чтобы красивее смотрелось на фронте
         this.newSearch = Object.assign({}, search);
@@ -1356,12 +1382,12 @@ export default {
 
       this.tockens.forEach((item, i) => {
         // затем пробегаемся по все мтокенам
-        if (item.name == this.tempTocken.name) {
+        if (item.name === this.tempTocken.name) {
           // и смотрим есть ли у нас такой токен
           j = i;
         }
       });
-      if (j != -1 || Number.isInteger(index)) {
+      if (j !== -1 || Number.isInteger(index)) {
         // если токен уже есть
         let height = this.$el
           .querySelector('.block-tocken')
@@ -1426,10 +1452,10 @@ export default {
     yesSearch: function () {
       // кнопка согласия на обновления если ИС или токен уже существует
       let elem =
-        event.target.nodeName.toLowerCase() != 'button'
+        event.target.nodeName.toLowerCase() !== 'button'
           ? event.target.parentElement
           : event.target; // сперва берем родителя кнопки, и если не получилось поймать кнопку, то еще выше уровнеь берем
-      if (elem.getAttribute('tool') == 'search') {
+      if (elem.getAttribute('tool') === 'search') {
         // если это окно ИС
         this.$store.commit('setSearch', {
           search: this.newSearch,
@@ -1440,7 +1466,7 @@ export default {
           .querySelector('.warning-block')
           .classList.remove('warning-block-show'); // убираем окно с предпреждением
         this.openSearch();
-      } else if (elem.getAttribute('tool') == 'tocken') {
+      } else if (elem.getAttribute('tool') === 'tocken') {
         // если это токен - собственно тоже самое
         const id = this.index;
         const newName = this.tempTocken.name;
@@ -1554,13 +1580,13 @@ export default {
     dragTool: function (event) {
       // функция для перетаскивания нового элемнета на полотно (остальную область, ну вы поняли короче)
 
-      if (event.which != 1) {
+      if (event.which !== 1) {
         // если пошло что-то не так
         return; // то прекращаем функцию
       }
       let parent = '';
       let elem = '';
-      if (event.target.nodeName != 'div') {
+      if (event.target.nodeName !== 'div') {
         // если мы ухватились не за div
         elem = event.target;
         while (!elem.classList.contains('tool-one')) {
@@ -1579,7 +1605,7 @@ export default {
       let avatar = parent.cloneNode(true); // дальше мы создаем как бы клон нашего элемнета
       document.body.appendChild(avatar); // и его уже добовляем в body
       avatar.classList.add('avatar'); // даем ему класс
-      avatar.style.zIndex = 3; // делаем его выше всех
+      avatar.style.zIndex = 99; // делаем его выше всех
       avatar.style.position = 'absolute'; // и относительно позиионируем
 
       document.onmousemove = (event) => {
@@ -1677,7 +1703,7 @@ export default {
       if (size.left + size.width > clientWidth) {
         result.left = clientWidth - size.width - 20;
       } else {
-        if (result.left == 0) {
+        if (result.left === 0) {
           result.left = size.left;
         }
       }
@@ -1694,14 +1720,14 @@ export default {
     openSettings: function () {
       this.$emit('openSettings');
 
-      if (this.colorGear == 'controlsActive') {
+      if (this.colorGear === 'controlsActive') {
         this.colorGear = 'controls';
       } else {
         this.colorGear = 'controlsActive';
       }
     },
     openExim: function () {
-      if (this.colorExim == 'controlsActive') {
+      if (this.colorExim === 'controlsActive') {
         this.colorExim = 'controls';
       } else {
         this.colorExim = 'controlsActive';
@@ -1709,15 +1735,15 @@ export default {
       this.openexim = !this.openexim;
     },
     setEvents: function () {
-      if (this.textarea_event != null && this.textarea_event != '') {
+      if (this.textarea_event !== null && this.textarea_event !== '') {
         let events = this.textarea_event.split('\n');
         let reg, body, bodyArray, element, doing, originItem;
 
-        if (events.length != 0) {
+        if (events.length !== 0) {
           events.forEach((item) => {
             originItem = item;
             item = item.replace(/\s/g, '');
-            if (item != '') {
+            if (item !== '') {
               reg = new RegExp(/^[\s+]?[\w]+\(/, 'g');
               this.$set(
                 this.event,
@@ -1729,13 +1755,13 @@ export default {
               body = body.slice(1, body.length - 1);
               bodyArray = body.split(',');
               bodyArray.forEach((elem, i) => {
-                if (elem.indexOf('(') != -1) {
+                if (elem.indexOf('(') !== -1) {
                   element = bodyArray.splice(0, i);
                 }
               });
 
-              if (this.event.event == 'OnDataCompare') {
-                if (element.length > 2 && element[1].indexOf('[') == -1) {
+              if (this.event.event === 'OnDataCompare') {
+                if (element.length > 2 && element[1].indexOf('[') === -1) {
                   this.$set(this.event, 'compare', element[0]);
                   this.$set(this.event, 'column', element[1]);
                   this.$set(
@@ -1751,7 +1777,7 @@ export default {
                     element.splice(1, element.length - 1).join(',')
                   );
                 }
-              } else if (this.event.event == 'OnTokenCompare') {
+              } else if (this.event.event === 'OnTokenCompare') {
                 this.$set(this.event, 'compare', element[0]);
                 this.$set(this.event, 'token', element[1]);
                 this.$set(
@@ -1759,13 +1785,13 @@ export default {
                   'tokenval',
                   element.splice(2, element.length - 1).join(',')
                 );
-              } else if (this.event.event == 'onValueCompare') {
-                if (element.length == 2) {
+              } else if (this.event.event === 'onValueCompare') {
+                if (element.length === 2) {
                   this.$set(this.event, 'treshold', element[0]);
                   this.$set(this.event, 'color', element[1]);
                 } else {
                   for (let i = 0; i < element.length; i++) {
-                    if (element[i].indexOf(']') != -1) {
+                    if (element[i].indexOf(']') !== -1) {
                       this.$set(
                         this.event,
                         'treshold',
@@ -1783,10 +1809,10 @@ export default {
               } else {
                 this.$set(this.event, 'element', element[0]); //click
                 if (element[1]) {
-                  if (element[1].indexOf('[') != -1) {
+                  if (element[1].indexOf('[') !== -1) {
                     let j = -1;
                     element.forEach((item, i) => {
-                      if (item.indexOf(']') != -1) {
+                      if (item.indexOf(']') !== -1) {
                         j = i;
                       }
                     });
@@ -1806,13 +1832,13 @@ export default {
               doing = reg.exec(body)[0];
               doing = doing.split('(');
               this.$set(this.event, 'action', doing[0]);
-              if (doing[0].toLowerCase() == 'set'.toLowerCase()) {
+              if (doing[0].toLowerCase() === 'set'.toLowerCase()) {
                 doing = doing[1].slice(0, doing[1].length - 1).split(',');
 
                 this.$set(this.event, 'target', doing[0]);
                 doing.splice(0, 1);
                 doing = doing.join(',');
-                if (doing.indexOf('[') != -1 && doing.indexOf(']') != -1) {
+                if (doing.indexOf('[') !== -1 && doing.indexOf(']') !== -1) {
                   doing = doing.match(/[^\[]+(?=\])/g);
                 } else {
                   doing = doing.split(',');
@@ -1829,13 +1855,13 @@ export default {
                     this.$set(this.event, 'value', ['']);
                   }
                 }
-              } else if (doing[0].toLowerCase() == 'go'.toLowerCase()) {
+              } else if (doing[0].toLowerCase() === 'go'.toLowerCase()) {
                 ///go
                 doing = doing[1].slice(0, doing[1].length - 1).split(',');
                 this.$set(this.event, 'target', doing[0]);
 
                 let prop, value;
-                if (doing[1].indexOf('[') != -1) {
+                if (doing[1].indexOf('[') !== -1) {
                   doing.splice(0, 1);
                   doing = doing.join(',');
                   doing = doing.match(/[^\[]+(?=\])/g);
@@ -1848,7 +1874,7 @@ export default {
                 this.$set(this.event, 'prop', prop);
                 this.$set(this.event, 'tab', doing[2]);
                 this.$set(this.event, 'value', value);
-              } else if (doing[0].toLowerCase() == 'open'.toLowerCase()) {
+              } else if (doing[0].toLowerCase() === 'open'.toLowerCase()) {
                 //open
                 doing = doing[1].slice(0, doing[1].length - 1).split(',');
 
@@ -1861,14 +1887,14 @@ export default {
 
                 this.$set(this.event, 'header', doing[5]);
               } else if (
-                doing[0].toLowerCase() == 'changeReport'.toLowerCase()
+                doing[0].toLowerCase() === 'changeReport'.toLowerCase()
               ) {
                 // changeReport
 
                 doing = originItem.split(doing[0])[1];
                 doing = doing.replace(/\(/g, '').replace(/\)/g, '').split(',');
                 this.$set(this.event, 'sid', doing[0]);
-                if (doing[1].indexOf('[') != -1) {
+                if (doing[1].indexOf('[') !== -1) {
                   doing.splice(0, 1);
                   let files = doing.map((item) => {
                     return item.replace('[', '').replace(']', '');
@@ -1878,7 +1904,7 @@ export default {
                   this.$set(this.event, 'file', [doing[1]]);
                 }
               } else if (
-                doing[0].toLowerCase() == 'exportSearch'.toLowerCase()
+                doing[0].toLowerCase() === 'exportSearch'.toLowerCase()
               ) {
                 // changeReport
 
@@ -1918,7 +1944,7 @@ export default {
     },
 
     changeColor: function () {
-      if (document.querySelectorAll('.v-menu__content').length != 0) {
+      if (document.querySelectorAll('.v-menu__content').length !== 0) {
         document.querySelectorAll('.v-menu__content').forEach((item) => {
           item.style.boxShadow = `0 5px 5px -3px ${this.theme.border},0 8px 10px 1px ${this.theme.border},0 3px 14px 2px ${this.theme.border}`;
           item.style.background = this.theme.back;
@@ -1935,7 +1961,7 @@ export default {
       });
       response.then((res) => {
         this.errorSave = true;
-        if (res.status == 200) {
+        if (res.status === 200) {
           this.colorErrorSave = this.theme.controls;
           this.msgErrorSave = 'Дашборд сохранен';
           this.$store.auth.getters.putLog(