+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - fixed node label name in YGraph
 - fixed the initial position of dash
+- fixed dash z index
+- fixed tooltip z-index
+
 
 ## [1.81.0]